+++ History.markdown
@@ -4,6 +4,11 @@
 
   * Test against Ruby 2.5 (#57)
   * Rely on rubocop-jekyll (#62)
+  * chore(deps): rubocop-jekyll 0.3 (#64)
+
+### Major Enhancements
+
+  * chore(deps): drop support for Liquid &lt; 4.0 (#66)
 
 ## 1.5.0 / 2017-12-03